loadbalancer: add shared logic for endpoints reflector
Extract some of the endpointslice reflector logic in a new package to
allow clustermesh to share some part of the code from the k8s reflector.
